refactor: split utils/auth.ts into modular structure (668→7 files)
## 変更内容
utils/auth.ts（668行）を責任範囲別に7つのモジュールに分割

### 新しい構造
```
src/utils/auth/
├── types.ts            (47行) - 型定義
├── guards.ts           (79行) - 認証ガード
├── audit-logger.ts     (44行) - 監査ログ
├── lockout-manager.ts  (208行) - ロックアウト管理
├── password-manager.ts (126行) - パスワード管理
├── session-manager.ts  (165行) - セッション管理
└── index.ts            (24行) - 統合エクスポート
```

### リファクタリング効果
- **保守性向上**: 各モジュールが単一責任を持つ
- **可読性向上**: 最大208行（推奨500行以内）
- **テスタビリティ**: モジュール単位でのテスト作成が容易
- **後方互換性**: `@/utils/auth`のインポートパスを維持

### 品質確認
- ✅ 型チェック成功
- ✅ ESLint成功
- ✅ 既存インポートパスの互換性維持

## Root Cause Analysis

**なぜこの問題が発生したか:**
- [x] 設計段階での考慮不足 - 単一ファイルに複数の責任が集約
- [x] 実装時のロジックミス - 段階的な機能追加で肥大化

**具体的な原因:**
認証機能の段階的な追加（セッション管理、ロックアウト、パスワード検証など）により、
単一ファイルが668行に肥大化し、保守性が低下していた。

**今後の予防策:**
- 新機能追加時は単一責任原則を遵守
- ファイル行数が500行を超える前にリファクタリング
- ESLintルール`max-lines`の設定検討

Closes #251
